Update index.html
more drama changing colors of the "next rec" element. we are not finding the element when the timer gets low. it is like the element class name changes or something else about the document structure. so no i am just gonna recolor the timer element, because that seems more reliable