Update index.html
updated autocapture to also capture images. Captures now consist of an object with a data field and image field. the data field contains what used to be the entirety of a capture, except now it is not stringified. however the entire captures array is  still stringified obviously. adding functions to capture image directly to clipboard and retrieve image by number. Change some function names for clarity. hope this works. untested