Update index.html
captureImageV0 is currently not in use. captureImage is now updated to collage properly. turns out the table heading row is duplicated in the original dom. one is for when the table is scrolled (this one is not part of the table element) and the other for when it is not scrolled (this one is part of the table element) I updated capture image to only capture the one which is a part of the table element. also made collaging happen before conversion to dataurl for efficiency