Update index.html
Added nonblocking alerts for captures in order to provide peace of mind that the autocapture is working. The alerts show number of rows and length of image url captured. Added alerts for indications of interruption by rec refresh, resulting in abort (when a capture is aborted, the aborted data is saved to local storage). also added alerts for worrisome errors during auto capture (these show up red on left side).