Update index.html
ok amoung other changes to reduce the amount of taffic to and from localStorage, captures are now stored in seperate fields of local storage. The prompt to retrieve now displays the numbers available and time captured, as 0 is no longer the most recent. The position at which a capture is stored changes in a modulo fashion from 0 to max captures minus one