Update index.html
Adding bookmarklets for capturing and retrieving a simpler summary to avoid ruffling feathers or overloading with unnecessary information. Implemented by adding a boolean input to the function that gets a summary from a dataobj. Also fixed a bug that counted rec grouping rows as recommendations when getting the total count of recommendations. Now the count filters according to class name of the row and also checks number of columns as a safety precaution in case class names change in the future.